requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.2
tqdm==4.66.4
pydantic==1.10.15
